optee-os: backport gcc10 compilation fix from upstream
Backport PR 3891 [1] from OP-TEE upstream repository. This PR solves
GCC10 compilation issue when new compiler option `outline-atomics` is
not implemented and causes the linker error.

In addition, above upstream PR solves issue with 'cc-option` macro,
which is necessary to be used to provide comptibility with GCC <10 to
disable the `outline-atomics` option, since it is present only in GCC10.

[1]: https://github.com/OP-TEE/optee_os/pull/3891

Signed-off-by: Andrey Zhizhikin <andrey.zhizhikin@leica-geosystems.com>